backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
gunicorn
redis                 # опційно
django-cors-headers>=4,<5
requests
celery[redis]
uvicorn[standard]>=0.23.0
gspread
google-auth
cryptography

